fix: moved automorphic files to math package
